Web Panel: Analogue Graphing Y-scale
Testing commenting out the min and max for the yAnalogue scale of the graph so that the y axis scale will expand to cover the range of the displayed values so that small ranges can actually be seen. Need to work out a better method for scales when multiple channels are being graphed with very different ranges.